fix: resolve action name mismatches in unified routers and harden type guards
- Fix bulk task actions: bulk_create, bulk_update, bulk_delete, bulk_complete
- Fix task-ops day_order action mismatch
- Fix project collaborators and move_to_parent action mismatches
- Fix subtask hierarchy action mismatch
- Remove unreachable get_collaborators from project ops router
- Add null/object checks to type guards (isFindDuplicatesArgs, isGetActivityArgs, etc.)
- Update server version from 0.11.0 to 1.0.0
- Extract formatTaskHierarchy to shared utility (src/utils/formatters.ts)
